Move ModelAnnotation::AnnotationBuilder under Annotation namespace
