+++ lib/annotate_rb/model_annotator/annotation.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module AnnotateRb
+  module ModelAnnotator
+    module Annotation
+      autoload :AnnotationBuilder, "annotate_rb/model_annotator/annotation/annotation_builder"
+    end
+  end
+end
